added some small but important stuff

+++ mainbody.sh
@@ -1,5 +1,10 @@
-
 #!/bin/bash
+
+#Makes sure directory to be backed up exists
+if [ -d $1 ];
+	then
+	return;
+	fi
 
 rm myJob*
 mkdir ~/Backups/
@@ -8,7 +13,7 @@ mkdir ~/Backups/$1/
 #Wrap submition task
 
 echo "#!/bin/bash
-#MSUB -N ajob
+#MSUB -N myJob
 #MSUB -A b1011
 #MSUB -q ligo
 #MSUB -l walltime=00:00:00:20
@@ -17,14 +22,14 @@ echo "#!/bin/bash
 
 cd \$PBS_O_WORKDIR
 
-#Finds recent backup to use in dif
-Lastback='ls -t ~/Backups/$1 | head -1
+#Finds recent backup to use in diff
+lastback=~/Backups/$1/$(ls -t ~/Backups/$1 | head -1)
 
-log()
+source log.sh
 
 cp $1 ~/Backups/$1/$(date +%F_%H.%M.%S)
 
-diskspace()" > submit;
+source diskspace.sh" > submit;
 
 msub submit;
 rm submit;
@@ -41,15 +46,8 @@ time=`cat myJob* | grep -Eo 'cput=.{0,8}' | tr -d 'cput='
 
 
 echo "Backup completed"
-echo "cpu time:", time
+echo "cpu time:"$time
 
-
-#cd .. | if [ -a FILE ] == False;
-#	then mkdir backups	
-#	else then a
-#	fi
-
-#cp $0  
 
 #Thought we should leave log file in backup folder
 mv myJob* ~/Backups/$1/$(ls -t | head -1)
